first try day 10

+++ help/help.h
@@ -19,7 +19,7 @@
 #define sza(x) ((int)x.size())
 
 #define PI 3.1415926535897932384626433832795l 
-const int MAX_N = 1e5 + 5;
+const double MAX_N = 1e5 + 5;
 const ll MOD = 1e9 + 7;
 const ll INF = 1e9;
 const ld EPS = 1e-9;
